requirements.txt
firecrawl-py
python-dotenv
pandas
openpyxl
google-generativeai